Add jQuery
jQuery is now being injected by Webpack too. In order to do so, you need
to require webpack so that we can use the ProvidePlugin. In the config
line for that plugin we need to map the function provided by jQuery ($
and jQuery) to the module jquery that we installed with npm.